src/api-app/requirements.txt
flask-bcrypt
Flask-Migrate
pyjwt
Flask-Script
flask_testing
flask-restful
flask-cors
flask-jwt-extended
gunicorn